writer: add a writer interface
Similar to the reader interface, the writer interface provides a generic
way to write values to a file. The input is generic, taking a file, some
length or number of values to write, and a value. The return value is a
result containing the number of values written on success and an error
message on failure.