writer: update signature of writer interface
Rather than take a length argument, we leave it up to implementations to
sort out whether or not they need to use the length of an argument value
to call underlying file system APIs. Additionally, we also leave it up
to implementations to determine whether or not the write should take
ownership of the value.

Updates tests and call sites accordingly.

Thanks to @hellerve for the suggestion!