reader: update fread call improve byte-reader perf
This change was motivated by the performance of the original byte-reader
implementation, which was quite slow since it read a file one byte at a
time. Using fread instead, we can net some performance gains.

This currently depends on a signature change to fread in Carp's core
library which isn't at HEAD yet, so that blocks this change. Assuming
that change merges into upstream, we need to also update the
string-reader implementation--this commit handles that as well.